% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_data.R
\docType{data}
\name{abc_list}
\alias{abc_list}
\alias{abc_vec}
\title{Simulated Sequence Data from a Variable Length Markov Chain (VLMC)}
\format{
\itemize{
\item \code{abc_vec}: A character vector of length 1000 representing a single sequence generated by the VLMC.
\item \code{abc_list}: A list of three character vectors, each of length 1000, representing independent sequences generated by the VLMC.
}

An object of class \code{character} of length 1000.
}
\source{
The data was generated using the \code{rvlmc} function in the R package
\pkg{bacontrees}.
}
\usage{
abc_list

abc_vec
}
\description{
The data objects \code{abc_vec} and \code{abc_list} contain sequences generated using a Variable Length Markov Chain (VLMC) with a specified context tree and transition probabilities.
}
\details{
The VLMC used to generate the data was based on an alphabet of three symbols (\code{a}, \code{b}, and \code{c}) with a context tree constructed to a maximal depth of 3. The active contexts and their corresponding transition probabilities are as follows:
\itemize{
\item \verb{*.a} (global context): \code{c(0.10, 0.20, 0.70)}
\item \verb{*.b}: \code{c(0.33, 0.33, 0.34)}
\item \verb{*.c.a}: \code{c(0.20, 0.10, 0.70)}
\item \verb{*.c.b}: \code{c(0.01, 0.98, 0.01)}
\item \verb{*.c.c}: \code{c(0.40, 0.40, 0.20)}
}

The sequences were generated using a seed value of \code{1} to ensure reproducibility.
}
\examples{
# Access the single sequence
data(abc_vec)
print(abc_vec)

# Access the list of sequences
data(abc_list)
print(abc_list)

}
\keyword{datasets}
