% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_vlmc.R
\name{fit_vlmc}
\alias{fit_vlmc}
\title{Fit a VLMC model via Context Algorithm}
\usage{
fit_vlmc(data, cutoff = 10, max_length = 6)
}
\arguments{
\item{data}{Either a vector with discrete data or a lista of vectors.}

\item{cutoff}{Cutoff value for the (log) likelihood ratio test statistic.}

\item{max_length}{Depth of the maximal tree considered.}
}
\description{
Fit a VLMC model via Context Algorithm
}
