% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContextTree.R
\name{plot.ContextTree}
\alias{plot.ContextTree}
\title{Plot method for ContextTree class}
\usage{
\method{plot}{ContextTree}(x, ..., activeOnly = TRUE)
}
\arguments{
\item{x}{A \code{ContextTree} object.}

\item{...}{Not used.}

\item{activeOnly}{logical value. If TRUE, only the nodes in the
active tree are plotted (including internal nodes).}
}
\description{
Plot method for ContextTree class
}
\details{
The edges of the tree have their width corresponding to the number
of occurrences of contexts matching the child node.

The plot is done using ggraph and can be further modified.
}
