% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baConTree.R
\name{baConTree}
\alias{baConTree}
\title{Bayesian Context Tree R6 class}
\description{
Bayesian Context Tree R6 class

Bayesian Context Tree R6 class
}
\section{Super class}{
\code{\link[bacontrees:ContextTree]{bacontrees::ContextTree}} -> \code{baConTree}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-baConTree-new}{\code{baConTree$new()}}
\item \href{#method-baConTree-setAllDirichletPars}{\code{baConTree$setAllDirichletPars()}}
\item \href{#method-baConTree-setContextPriorWeights}{\code{baConTree$setContextPriorWeights()}}
\item \href{#method-baConTree-clone}{\code{baConTree$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bacontrees" data-topic="ContextTree" data-id="activateMaximal"><a href='../../bacontrees/html/ContextTree.html#method-ContextTree-activateMaximal'><code>bacontrees::ContextTree$activateMaximal()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bacontrees" data-topic="ContextTree" data-id="activateRoot"><a href='../../bacontrees/html/ContextTree.html#method-ContextTree-activateRoot'><code>bacontrees::ContextTree$activateRoot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bacontrees" data-topic="ContextTree" data-id="getActiveNodes"><a href='../../bacontrees/html/ContextTree.html#method-ContextTree-getActiveNodes'><code>bacontrees::ContextTree$getActiveNodes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bacontrees" data-topic="ContextTree" data-id="getGrowableNodes"><a href='../../bacontrees/html/ContextTree.html#method-ContextTree-getGrowableNodes'><code>bacontrees::ContextTree$getGrowableNodes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bacontrees" data-topic="ContextTree" data-id="getLeaves"><a href='../../bacontrees/html/ContextTree.html#method-ContextTree-getLeaves'><code>bacontrees::ContextTree$getLeaves()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bacontrees" data-topic="ContextTree" data-id="getParentNode"><a href='../../bacontrees/html/ContextTree.html#method-ContextTree-getParentNode'><code>bacontrees::ContextTree$getParentNode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bacontrees" data-topic="ContextTree" data-id="getSiblingNodes"><a href='../../bacontrees/html/ContextTree.html#method-ContextTree-getSiblingNodes'><code>bacontrees::ContextTree$getSiblingNodes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bacontrees" data-topic="ContextTree" data-id="growActive"><a href='../../bacontrees/html/ContextTree.html#method-ContextTree-growActive'><code>bacontrees::ContextTree$growActive()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bacontrees" data-topic="ContextTree" data-id="nodeExists"><a href='../../bacontrees/html/ContextTree.html#method-ContextTree-nodeExists'><code>bacontrees::ContextTree$nodeExists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bacontrees" data-topic="ContextTree" data-id="print"><a href='../../bacontrees/html/ContextTree.html#method-ContextTree-print'><code>bacontrees::ContextTree$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bacontrees" data-topic="ContextTree" data-id="root"><a href='../../bacontrees/html/ContextTree.html#method-ContextTree-root'><code>bacontrees::ContextTree$root()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bacontrees" data-topic="ContextTree" data-id="setData"><a href='../../bacontrees/html/ContextTree.html#method-ContextTree-setData'><code>bacontrees::ContextTree$setData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bacontrees" data-topic="ContextTree" data-id="validate"><a href='../../bacontrees/html/ContextTree.html#method-ContextTree-validate'><code>bacontrees::ContextTree$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baConTree-new"></a>}}
\if{latex}{\out{\hypertarget{method-baConTree-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baConTree$new(data, maximalDepth = 5, active = "root")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Either a vector with discrete data or a lista of vectors.}

\item{\code{maximalDepth}}{Depth of the maximal tree considered.}

\item{\code{active}}{Either "root" or "maximal" to indicate which nodes
should be initialized as active.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baConTree-setAllDirichletPars"></a>}}
\if{latex}{\out{\hypertarget{method-baConTree-setAllDirichletPars}{}}}
\subsection{Method \code{setAllDirichletPars()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baConTree$setAllDirichletPars(alpha)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{Hyperparameter considered for the Dirichlet prior distribution
of probabilities.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baConTree-setContextPriorWeights"></a>}}
\if{latex}{\out{\hypertarget{method-baConTree-setContextPriorWeights}{}}}
\subsection{Method \code{setContextPriorWeights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baConTree$setContextPriorWeights(fn)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fn}}{A function to be evaluated at each node that returns
its weight in the prior distribution.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baConTree-clone"></a>}}
\if{latex}{\out{\hypertarget{method-baConTree-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baConTree$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
