% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacontrees-package.R
\docType{package}
\name{bacontrees-package}
\alias{bacontrees}
\alias{bacontrees-package}
\title{bacontrees: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Victor Freguglia \email{victorfreguglia@gmail.com} (\href{https://orcid.org/0000-0002-6189-4453}{ORCID})

}
\keyword{internal}
