% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis_vlmc.R
\name{metropolis_vlmc}
\alias{metropolis_vlmc}
\title{Metropolis-Hastings for Context Trees}
\usage{
metropolis_vlmc(
  data,
  n_steps,
  max_depth = 6,
  alpha = 0.001,
  context_weights = function(node) 0,
  burnin = 100,
  thin = 1,
  initial_root = TRUE
)
}
\arguments{
\item{data}{Either a vector with discrete data or a lista of vectors.}

\item{n_steps}{Number of steps to run the Metropolis-Hastings algorithm for.}

\item{max_depth}{Maximum depth considered for the context tree (using high
values may demand too much memory).}

\item{alpha}{Hyper parameters for the Dirichlet prior distribution used
for transition probabilities.}

\item{context_weights}{A function of a TreeNode that returns the contribution
of a given context in the tree prior distribution (in log scale).
See \link[bacontrees]{TreeNode} for details.}

\item{burnin}{Number of initial iterations to exclude from posterior
summarization results.}

\item{thin}{Number of iterations between observations used in the
posterior distribution summarization.}

\item{initial_root}{Logical value indicating whether the initial tree
should contain only the root node. If set to \code{FALSE}, the maximal tree
of the specified depth is used.}
}
\value{
A \code{metropolis_vlmc} object containing attributes:
\itemize{
\item \code{df}: A table containing the set of contexts, tree code, counts and posterior probabilities
considering the specified burnin and thin values.
\item \code{codes}: A list containing the contexts for the trees corresponding to each tree code.
\item \code{chain}: The complete chain generated, represented by tree codes.
}
}
\description{
Generates a Markov Chain of Context Trees using the Metropolis-Hastings
algorithm given some sequence data set.
}
