% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvlmc.R
\name{rvlmc}
\alias{rvlmc}
\title{Generate a sequence based on a Variable Length Markov Chain (VLMC)}
\usage{
rvlmc(n, alphabet, context_list, context_probs)
}
\arguments{
\item{n}{An integer specifying the length of the sequence to generate.}

\item{alphabet}{A character vector representing the set of symbols used in the sequence.}

\item{context_list}{A character vector specifying the list of contexts in the VLMC. Each context should be a string where symbols are separated by dots (e.g., \code{"a.b.c"}).}

\item{context_probs}{A list of probability vectors corresponding to each context in \code{context_list}. Each vector specifies the probabilities of sampling each symbol in \code{alphabet} given the context.}
}
\value{
A character vector of length \code{n} representing the generated sequence.
}
\description{
This function simulates a sequence using a Variable Length Markov Chain (VLMC) model.
}
\examples{
# Define parameters for the VLMC
n <- 1000
alphabet <- c("a", "b", "c")
context_list <- c("*.a", "*.b", "*.c.a", "*.c.b", "*.c.c")
context_probs <- list(
  c(0.10, 0.20, 0.70), #.a
  c(0.33, 0.33, 0.34), #.b
  c(0.20, 0.10, 0.70), #.c.a
  c(0.01, 0.98, 0.01), #.c.b
  c(0.40, 0.40, 0.20)  #.c.c
)

# Generate a sequence
sequence <- rvlmc(n, alphabet, context_list, context_probs)
print(sequence)

}
