% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeFromContexts.R
\name{treeFromContexts}
\alias{treeFromContexts}
\title{Create ContextTree object from a vector of contexts}
\usage{
treeFromContexts(contexts)
}
\arguments{
\item{contexts}{Either a string with contexts separated by \code{','} or
a vector of strings representing contexts.}
}
\value{
A \code{ContextTree} object with maximal length equal to the largest
depth among specified contexts and active tree equal to the Context Tree
specified by the contexts.
}
\description{
Create ContextTree object from a vector of contexts
}
\examples{
treeFromContexts(c("*.0", "*.1.0", "*.1.1"))
treeFromContexts("{*.a, *.b, *.c.a, *.c.b, *.c.c}")


}
